fix: fail closed on impossible report transitions
